Move service related code in client/model to client/services/models. Rename populate_services.js to api_load_has_services.js
