+++ client/src/panels/panel_declarations/services/services_channels.js
@@ -16,7 +16,7 @@ import {
   useServiceSummaryGov,
   useServiceSummaryOrg,
   useServiceSummaryProgram,
-} from "src/models/services_queries";
+} from "src/models/services/services_queries";
 
 import { infobase_colors } from "src/core/color_schemes";
 import { formats } from "src/core/format";
