+++ client/src/panels/panel_declarations/services/services_digital_status.js
@@ -16,7 +16,7 @@ import {
   useServiceSummaryGov,
   useServiceSummaryOrg,
   useServiceSummaryProgram,
-} from "src/models/services_queries";
+} from "src/models/services/services_queries";
 
 import { is_a11y_mode, lang } from "src/core/injected_build_constants";
