+++ client/src/panels/panel_declarations/services/services_standards.js
@@ -13,7 +13,7 @@ import {
   useServiceSummaryGov,
   useServiceSummaryOrg,
   useServiceSummaryProgram,
-} from "src/models/services_queries";
+} from "src/models/services/services_queries";
 
 import { is_a11y_mode } from "src/core/injected_build_constants";
