+++ client/src/panels/panel_declarations/services/subject_offering_services.js
@@ -14,7 +14,7 @@ import {
 import {
   useServiceSummaryGov,
   useServiceSummaryOrg,
-} from "src/models/services_queries";
+} from "src/models/services/services_queries";
 
 import { Dept, Program } from "src/models/subjects";
